misc/HtmlTestSuite/results/test-1045.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><iframe id="i"/>

<script>
async_test(t =&gt; {
  let starting_history_length = history.length;
  let iframe_url = (new URL("/common/blank.html", location.href)).href;
  i.src = iframe_url;

  history.pushState("a", "", "#a");
  assert_equals(history.length, starting_history_length + 1, "First history length");

  i.onload = t.step_func(() =&gt; {
    assert_equals(history.length, starting_history_length + 1, "Second history length");
    assert_equals(i.contentWindow.location.href, iframe_url);
    assert_equals(location.hash, "#a");
    history.back();
    // Wait a while for a back navigation. Since all of the possible outcomes
    // are either same-document or navigating to about:blank, this doesn't need
    // to wait terribly long.
    t.step_timeout(t.step_func_done(() =&gt; {
      assert_equals(location.hash, "", "top frame should have navigated back");
      assert_equals(i.contentWindow.location.href, iframe_url, "iframe should not have navigated");
    }), 100);
  });
}, "pushState() in parent while child is doing initial navigation, then go back");
</script>
</body>
</html>
